// packages-test-3/ai-chat/src/stream/tool-types.ts
export type ToolExecuteFunction<TArgs, TResult> = (
  args: TArgs,
  context: ToolExecutionContext,
) => TResult | Promise<TResult>;

export type ToolStreamCallFunction<
  TArgs extends Record<string, unknown> = Record<string, unknown>,
  TResult = unknown,
> = (
  reader: ToolCallReader<TArgs, TResult>,
  context: ToolExecutionContext,
) => void;

type OnSchemaValidationErrorFunction<TResult> = ToolExecuteFunction<
  unknown,
  TResult
>;

type ToolBase<
  TArgs extends Record<string, unknown> = Record<string, unknown>,
  TResult = unknown,
> = {
  /**
   * @deprecated Experimental, API may change.
   */
  streamCall?: ToolStreamCallFunction<TArgs, TResult>;
};

type BackendTool<
  TArgs extends Record<string, unknown> = Record<string, unknown>,
  TResult = unknown,
> = ToolBase<TArgs, TResult> & {
  type?: "backend" | undefined;

  description?: undefined;
  parameters?: undefined;
  disabled?: undefined;
  execute?: undefined;
  experimental_onSchemaValidationError?: undefined;
};

type FrontendTool<
  TArgs extends Record<string, unknown> = Record<string, unknown>,
  TResult = unknown,
> = ToolBase<TArgs, TResult> & {
  type?: "frontend" | undefined;

  description?: string | undefined;
  parameters: StandardSchemaV1<TArgs> | JSONSchema7;
  disabled?: boolean;
  execute?: ToolExecuteFunction<TArgs, TResult>;
  experimental_onSchemaValidationError?: OnSchemaValidationErrorFunction<TResult>;
};

type HumanTool<
  TArgs extends Record<string, unknown> = Record<string, unknown>,
  TResult = unknown,
> = ToolBase<TArgs, TResult> & {
  type?: "human" | undefined;

  description?: string | undefined;
  parameters: StandardSchemaV1<TArgs> | JSONSchema7;
  disabled?: boolean;
  execute?: undefined;
  experimental_onSchemaValidationError?: undefined;
};

export type Tool<
  TArgs extends Record<string, unknown> = Record<string, unknown>,
  TResult = unknown,
> =
  | FrontendTool<TArgs, TResult>
  | BackendTool<TArgs, TResult>
  | HumanTool<TArgs, TResult>;